Move isScriptureEqual & merge set to their own files. Tweak imports/exports
