+++ src/components/product/product.module.ts
@@ -1,7 +1,7 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { AuthorizationModule } from '../authorization/authorization.module';
 import { FileModule } from '../file/file.module';
-import { ScriptureModule } from '../scripture/scripture.module';
+import { ScriptureModule } from '../scripture';
 import { StoryModule } from '../story/story.module';
 import * as handlers from './handlers';
 import * as migrations from './migrations';
