+++ src/components/scripture/index.ts
@@ -1,3 +1,10 @@
 export * from './dto';
+export * from './books';
+export * from './merge-to-minimal-set';
+export * from './parser';
+export * from './is-equal';
+export * from './labels';
+export * from './verse-equivalents';
 export * from './scripture-reference.repository';
 export * from './scripture-reference.service';
+export * from './scripture.module';
